requirements.txt
async-lru==1.0.2
dnspython==2.1.0
mongoengine==0.23.0
Pillow==8.2.0
pyaes==1.6.1
pymongo==3.11.3
Pyrogram==1.2.6
PySocks==1.7.1
TgCrypto==1.2.2
